Invert API layering: marks.py is the declarative mark core
The nine mark-builder implementations (line/area/scatter/histogram/hist/
bar/column/heatmap and the private _bar_like) move verbatim from figure.py
into the new declarative core module marks.py. Figure binds them in its
class body (Figure.scatter is marks.scatter), so the fluent API is now a
veneer over the declarative implementations: one body, one signature, one
set of defaults for both dialects, with zero added frames or allocations —
payload output is byte-identical (verified against a 19-case fluent+
declarative fingerprint matrix).

Gate updates: fastcharts.marks joins HEAVY_MODULES (test_import),
HEAVY_FASTCHARTS_IMPORTS (check_public_api), and REQUIRED_FILES
(verify_wheel — figure.py markers now assert the binding surface and
marks.py carries the def markers, with a new stale-marks negative test).
